feat: Fixing problems with generated code for golang and adding sample manifest for testing. (#22)
<!-- Please use this template for your pull request. -->
<!-- Please use the sections that you need and delete other sections -->

## This PR
<!-- add the description of the PR here -->
Several small changes:
- Add sample manifest
- Fix integer parsing and generating for golang
- Small fixes for missing defaults in switches


### Related Issues
No issue - I've just tried to use the code to generate accessors

---------

+++ src/providers/providers.go
@@ -6,5 +6,5 @@ import (
 
 type BooleanProvider func(ctx context.Context) (bool, error)
 type FloatProvider func(ctx context.Context) (float64, error)
-type IntProvider func(ctx context.Context) (int, error)
+type IntProvider func(ctx context.Context) (int64, error)
 type StringProvider func(ctx context.Context) (string, error)
